Rename and move builder scripts file
